+++ src/app/admin/units/unit/unit-list/unit-list.component.ts
@@ -12,6 +12,7 @@ import { UnitsService } from 'src/app/admin/units/units.service';
 import { GlobalService } from 'src/app/services/global.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { LocalService } from 'src/app/services/local.service';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-unit-list',
@@ -39,6 +40,7 @@ export class UnitListComponent {
                 private router: Router,
                 private global: GlobalService,
                 private localStore: LocalService,
+                public translateService: TranslateService,
                 public snackBar: MatSnackBar){
         this.settings = this.appSettings.settings;
     }
@@ -50,7 +52,7 @@ export class UnitListComponent {
         if (this.localStore.getItem('token') && this.localStore.getItem('account')) {
           this.global.me = JSON.parse(this.localStore.getItem('account'));
 
-          //this.prepareMsgLanguage() ;  //  for translation
+          this.prepareMsgLanguage() ;  //  for translation
           this.getUnits();
 
 
@@ -66,7 +68,9 @@ export class UnitListComponent {
     }
 // ===========================================================
 public removeUnit(unitId: number){
-  this.unitsService.deleteUnit(unitId).subscribe(user => this.getUnits());
+  if(confirm(this.askToDeletedMsg)) {
+    this.unitsService.deleteUnit(unitId).subscribe(user => this.getUnits());
+  }
 }
 // ===========================================================
     public onPageChanged(event){
@@ -78,7 +82,20 @@ public removeUnit(unitId: number){
 gotoNewUnit(){
   this.router.navigate(["/units/unit-add/"]);
 }
+// ====================================================================
+errorRetrieveMsg ;
+askToDeletedMsg ;
+deletedMsg ;
+prepareMsgLanguage(){
+  this.translateService.get('MESSAGE.RETRIEVE_ERROR', ).subscribe((res: string) => {
+    this.errorRetrieveMsg = res ;  });
+    this.translateService.get('MESSAGE.SURE_DELETE', ).subscribe((res: string) => {
+      this.askToDeletedMsg = res ;  });
+      this.translateService.get('MESSAGE.DELETED', ).subscribe((res: string) => {
+        this.deletedMsg = res ;  });
 
+}
+// ====================================================================
 }
 
 // this.snackBar.open(this.deletedMsg, '×', { panelClass: 'success', verticalPosition: 'top', duration: 3000 });
